==> 01 corpus acquisition/04_create_annotation_dataset.ipynb <==
# Notebook: Filter Sentences from Reviews to create Dataset for Annotation

## Packages

from bs4 import BeautifulSoup
from langdetect import detect, DetectorFactory
import pandas as pd
import nltk
import uuid
from nltk.tokenize import sent_tokenize

## Constants

REVIEWS_PATH = "reviews_dataset/reviews.csv"
RANDOM_STATE = 43
LOAD_1000_ADDITIONAL_SENTENCES = True

## Settings

DetectorFactory.seed = RANDOM_STATE

nltk.download('punkt')

## Code

### Load Dataset

df_reviews = pd.read_csv(REVIEWS_PATH)

len(df_reviews)

# Höchstes Datum
max_date = df_reviews['date'].max()

# Niedrigstes Datum
min_date = df_reviews['date'].min()

print("Höchstes Datum:", max_date)
print("Niedrigstes Datum:", min_date)

### Filter Sentences

df_reviews_sentences = pd.DataFrame(columns=list(df_reviews.columns) + ['sentence_idx', "id"])

for idx, row in df_reviews.iterrows():
    sentences = sent_tokenize(row['text'], language="german")
    sentence_index = 0
    for sentence in sentences:
        if len(sentence) > 1:
            new_row = row.copy() 
            new_row['text'] = sentence 
            new_row['id'] = str(uuid.uuid4())
            new_row['sentence_idx'] = sentence_index
            df_reviews_sentences.loc[len(df_reviews_sentences)] = new_row
            sentence_index += 1

df_reviews_sentences

### Filter Non-German Sentences

Even if the language code = "de", we have observed that some sentences are in other languagesthan german. These are excluded as well with the help of Google's language-detection.

def detect_language(text):
    try:
        return detect(text)
    except:
        return 'Unknown'

df_reviews_sentences['detected_language'] = df_reviews_sentences['text'].apply(detect_language)
df_reviews_sentences[df_reviews_sentences['detected_language'] != 'de']

df_reviews_sentences = df_reviews_sentences[df_reviews_sentences['detected_language'] == 'de']

df_reviews_sentences

### Balancing

existing_sentences_path = 'reviews_dataset/balanced_reviews_sentences_4000.csv'

if LOAD_1000_ADDITIONAL_SENTENCES:
    df_4000 = pd.read_csv(existing_sentences_path)
    df_reviews_sentences = df_reviews_sentences[~df_reviews_sentences.set_index(['review_id', 'restaurant_id', 'sentence_idx']).index.isin(
    df_reviews_sentences.set_index(['review_id', 'restaurant_id', 'sentence_idx']).index.intersection(
        df_4000.set_index(['review_id', 'restaurant_id', 'sentence_idx']).index
      )
    )]

df_balanced_reviews_sentences = df_reviews_sentences.groupby(['rating'], group_keys=False).apply(lambda x: x.sample(200, replace=False, random_state=RANDOM_STATE))
df_balanced_reviews_sentences.reset_index(drop=True, inplace=True)

### Randomisierung

df_balanced_reviews_sentences = df_balanced_reviews_sentences.sample(frac=1, random_state=RANDOM_STATE)  # 'frac=1' mischt den gesamten DataFrame
df_balanced_reviews_sentences.reset_index(drop=True, inplace=True)

### Check If Sentences are Unique

df_balanced_reviews_sentences.groupby(["review_id", "sentence_idx"]).size().reset_index(name="count")

### Alternate examples based on rating

df_balanced_reviews_sentences["rating"].value_counts()

dfs = [df_balanced_reviews_sentences[df_balanced_reviews_sentences["rating"] == i] for i in range(1, 6)]
df_alternating = pd.DataFrame(columns=df_balanced_reviews_sentences.columns)

max_rows = max(len(df) for df in dfs)
for i in range(max_rows):
    for k in range(5):
        if i < len(dfs[k]):
            df_alternating = pd.concat([df_alternating, dfs[k].iloc[[i]]])

df_balanced_reviews_sentences = df_alternating.reset_index(drop=True)

df_balanced_reviews_sentences

### Store as .csv 

if LOAD_1000_ADDITIONAL_SENTENCES == False:
   df_reviews_sentences.to_csv("reviews_dataset/reviews_sentences.csv")

df_balanced_reviews_sentences.to_csv("reviews_dataset/balanced_reviews_sentences.csv")

df_balanced_reviews_sentences

df_balanced_reviews_sentences["rating"].value_counts()
