==> 01 corpus acquisition/05_dataset_statistics.ipynb <==
# Notebook: View Dataset Statistics

This notebook is used to explore a given dataset.

## Packages

import pandas as pd
import tiktoken
import re

## Constants

DATASET_PATH = "reviews_dataset/balanced_reviews_sentences.csv"
LLM_MODEL_NAME = "gpt-3.5-turbo"

## Code

### General Statistics

df = pd.read_csv(DATASET_PATH)
df

rating_counts = df['rating'].value_counts()
rating_counts

rating_counts = df['city'].value_counts()
rating_counts

rating_counts = df['restaurant_name'].value_counts()
rating_counts

#df.groupby(['sentence_idx', 'review_id']).size().reset_index(name='count')

def count_words(text):
    words = re.findall(r'\w+', text)
    return len(words)

df['word_count'] = df['text'].apply(count_words)

# Standarddeviation
df['word_count'].std()


# Mean
df['word_count'].mean()

# Median
df['word_count'].median()

# Min, Max
df['word_count'].min(), df['word_count'].max()

# Source: https://github.com/openai/openai-cookbook/blob/main/examples/How_to_count_tokens_with_tiktoken.ipynb
def num_tokens_from_string(string: str, model_name: str) -> int:
    encoding = tiktoken.encoding_for_model(model_name)
    num_tokens = len(encoding.encode(string))
    return num_tokens

df['llm_token_count'] = df['text'].apply(num_tokens_from_string, model_name=LLM_MODEL_NAME)

df['llm_token_count'].min(), df['llm_token_count'].max()

top_10_highest_tokens = df.nlargest(10, 'llm_token_count')
top_10_highest_tokens


### Count Tokens (for LLM)

df['llm_token_count'] = df['text'].apply(lambda x: num_tokens_from_string(x, LLM_MODEL_NAME))

pd.set_option('display.max_colwidth', None)

max_token_row = df[df['llm_token_count'] == df['llm_token_count'].max()]
print("Zeile mit den meisten Tokens:")
max_token_row#["llm_token_count"].values[0]

with open("../prompt_template.txt", "r", encoding="utf-8") as file:
    prompt_template = file.read()

template_size = num_tokens_from_string(prompt_template, LLM_MODEL_NAME)
aspect_term_tag_size = num_tokens_from_string('<aspect-term aspect="ambient/interior" polarity="negative"></aspect-term>', LLM_MODEL_NAME)

template_size + 10 * max_token_row["llm_token_count"].values[0] + aspect_term_tag_size * 20
