==> 01 corpus acquisition/06_explore_topics_using_bertopic.ipynb <==
# Notebook: Explore Addressed Topics

This notebook is used to explore the topics discussed in reviews previously discussed. 

## Packages

from sklearn.cluster import AgglomerativeClustering
from sklearn.cluster import KMeans
from nltk.corpus import stopwords
from bertopic import BERTopic
from hdbscan import HDBSCAN
from umap import UMAP
import pandas as pd
import spacy
import nltk
import re

## Constants

nltk.download('stopwords')

DATASET_PATH = "reviews_dataset/balanced_reviews_sentences.csv"
STOPWORDS = set(stopwords.words('german'))
N_TOPICS = 6

## Code

#spacy.cli.download("de_core_news_sm")

nlp = spacy.load('de_core_news_sm')

dataset = pd.read_csv(DATASET_PATH)

def clean_text(text):
    RE_WSPACE = re.compile(r"\s+", re.IGNORECASE)
    RE_ASCII = re.compile(r"[^A-Za-zÀ-ž ]", re.IGNORECASE)
    RE_SINGLECHAR = re.compile(r"\b[A-Za-zÀ-ž]\b", re.IGNORECASE)

    text = re.sub(r'@\w+', '', text)
    text = re.sub(RE_ASCII, " ", text)
    text = re.sub(RE_SINGLECHAR, " ", text)
    text = re.sub(RE_WSPACE, " ", text)
    return text.strip()

def lemmatize_remove_stopwords_text(text):
    doc = nlp(text)
    lemmatized_text = ' '.join([token.lemma_.lower() for token in doc if token.text.lower() not in STOPWORDS])
    return lemmatized_text

dataset["text"] = dataset["text"].apply(clean_text)
dataset["text"] = dataset["text"].apply(lemmatize_remove_stopwords_text)

dataset = dataset.text.to_list()

# Use default UMAP but add random state
umap_model = UMAP(n_neighbors=15, n_components=5, min_dist=0.0, metric='cosine', random_state=43)
model = BERTopic(language="german", nr_topics=N_TOPICS, umap_model=umap_model, top_n_words=20, embedding_model="paraphrase-multilingual-MiniLM-L12-v2")

topics, probabilities = model.fit_transform(dataset)
topics = model.reduce_outliers(dataset, topics, strategy="distributions")
model.update_topics(dataset, topics, top_n_words=20)

model.get_document_info(dataset)["Topic"].value_counts().reset_index()

topic_statistics = model.get_document_info(dataset)["Topic"].value_counts().reset_index()#.rename(columns={"index": "Topic", "Topic": "Frequency"}).sort_values("Topic").reset_index(drop=True)
topic_statistics = topic_statistics[topic_statistics["Topic"] != -1]
topic_statistics["Topic"] = topic_statistics["Topic"] + 1
topic_statistics

top_words = []
for topic in topic_statistics["Topic"]:
    words = model.get_topic(topic - 1)[:5]
    top_words.append(', '.join([word[0] for word in words]))
    
topic_statistics["Top Words"] = top_words

def format_frequency(frequency):
    return "{:,}".format(frequency)

# Anwendung der Funktion auf die Spalte "Frequency"
topic_statistics["count"] = topic_statistics["count"].apply(format_frequency)

topic_statistics
