==> 01 corpus acquisition/07_convert_restaurants_to_latex.ipynb <==
german_to_english_cities = {
    'augsburg': 'Augsburg',
    'berlin': 'Berlin',
    'bielefeld': 'Bielefeld',
    'bochum': 'Bochum',
    'bonn': 'Bonn',
    'bremen': 'Bremen',
    'dortmund': 'Dortmund',
    'dresden': 'Dresden',
    'duisburg': 'Duisburg',
    'düsseldorf': 'Düsseldorf',
    'essen': 'Essen',
    'frankfurt': 'Frankfurt',
    'hamburg': 'Hamburg',
    'hannover': 'Hannover',
    'karlsruhe': 'Karlsruhe',
    'köln': 'Cologne',
    'leipzig': 'Leipzig',
    'mannheim': 'Mannheim',
    'mönchengladbach': 'Mönchengladbach',
    'münchen': 'Munich',
    'münster': 'Münster',
    'nürnberg': 'Nuremberg',
    'stuttgart': 'Stuttgart',
    'wiesbaden': 'Wiesbaden',
    'wuppertal': 'Wuppertal'
}


def add_thousand_dots(n_sample):
    if isinstance(n_sample, str):
        if '.' in n_sample:
            integer_part, decimal_part = n_sample.split('.')
            formatted_integer_part = "{:,}".format(int(integer_part))
            result = f"{formatted_integer_part}.{decimal_part}"
        else:
            result = "{:,}".format(int(n_sample))
    elif isinstance(n_sample, np.float64):
        result = "{:,}".format(round(n_sample, 1))
    else:
        result = n_sample
    
    return result

import json

# Lade die JSON-Datei
with open('restaurants.json', 'r') as file:
    data = json.load(file)

# Teile den Datensatz in vier Teile auf
n_tables = 2
num_records = len(data)
chunk_size = num_records // n_tables

for idx in range(n_tables):
    start = idx * chunk_size
    end = (idx + 1) * chunk_size if idx < 3 else num_records
    chunk_data = data[start:end]

    # LaTeX-Tabelle erstellen
    latex_table = "\\begin{table}[H]\n"
    latex_table += "\\centering\n"
    latex_table += "\\tiny\n"
    latex_table += "\\rotatebox{90}{\n"
    latex_table += "\\begin{tabular}{llr}\n"
    latex_table += "\\hline\n"
    latex_table += "\\textbf{Restaurant URL} & \\textbf{City} & \\textbf{\\begin{tabular}[c]{@{}l@{}}\# User Reviews \\\ (As of October 31, 2023)\\end{tabular}} \\\\ \\hline\n"

    for restaurant in chunk_data:
        url = restaurant['url']
        city = german_to_english_cities[restaurant['city']]  # Nur erster Buchstabe der Stadt in Großbuchstaben
        n_ratings = add_thousand_dots(str(restaurant['n_ratings']))
        latex_url = "\\url{" + url.replace("_", "\\_") + "}"
        latex_table += f"{latex_url} & {city} & {n_ratings} \\\\ \n"

    latex_table += "\\hline\n"
    latex_table += "\\end{tabular}\n"
    latex_table += "}\n"
    latex_table += "\\end{table}"

    # Tabelle in eine LaTeX-Datei schreiben (z.B. restaurants_{idx}.tex)
    with open(f'restaurants_{idx}.tex', 'w') as latex_file:
        latex_file.write(latex_table)

    print(f"Die LaTeX-Tabelle wurde in 'restaurants_{idx}.tex' gespeichert.")
